+++ src/lib.rs
@@ -32,6 +32,7 @@ pub mod logger;
 pub mod metadata;
 pub mod mixer_audio;
 pub mod mixer_video;
+pub mod optuna;
 pub mod reader;
 pub mod reader_mp4;
 pub mod reader_webm;
