+++ src/subcommand_tune.rs
@@ -1,5 +1,4 @@
 use std::{
-    collections::BTreeMap,
     num::NonZeroUsize,
     path::{Path, PathBuf},
     process::{Command, Stdio},
@@ -8,7 +7,8 @@ use std::{
 use orfail::OrFail;
 
 use crate::{
-    json::{JsonNumber, JsonObject, JsonValue},
+    json::{JsonObject, JsonValue},
+    optuna::{Optuna, SearchSpace, TrialMetrics},
     subcommand_vmaf,
 };
 
@@ -119,7 +119,7 @@ pub fn run(mut args: noargs::RawArgs) -> noargs::Result<()> {
     }
 
     // 最初に optuna と vmaf コマンドが利用可能かどうかをチェックする
-    check_optuna_availability().or_fail()?;
+    Optuna::check_optuna_availability().or_fail()?;
     subcommand_vmaf::check_vmaf_availability().or_fail()?;
 
     // 必要なら tune_working_dir を作る
@@ -205,316 +205,6 @@ pub fn run(mut args: noargs::RawArgs) -> noargs::Result<()> {
     }
 
     Ok(())
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
-struct JsonObjectMemberPath(Vec<String>);
-
-impl JsonObjectMemberPath {
-    fn get<'a>(&self, mut value: &'a JsonValue) -> Option<&'a JsonValue> {
-        for name in &self.0 {
-            let JsonValue::Object(object) = value else {
-                return None;
-            };
-            value = object.get(name)?;
-        }
-        Some(value)
-    }
-
-    fn get_mut<'a>(&self, mut value: &'a mut JsonValue) -> Option<&'a mut JsonValue> {
-        for name in &self.0 {
-            let JsonValue::Object(object) = value else {
-                return None;
-            };
-            value = object.get_mut(name)?;
-        }
-        Some(value)
-    }
-}
-
-impl std::str::FromStr for JsonObjectMemberPath {
-    type Err = std::convert::Infallible;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Ok(Self(s.split('.').map(|s| s.to_owned()).collect()))
-    }
-}
-
-#[derive(Debug)]
-struct SearchSpace {
-    items: BTreeMap<JsonObjectMemberPath, SearchSpaceItem>,
-}
-
-impl SearchSpace {
-    fn new(root: nojson::RawJsonValue<'_, '_>) -> Result<Self, nojson::JsonParseError> {
-        let mut items = BTreeMap::new();
-        for (key, value) in root.to_object()? {
-            let path = JsonObjectMemberPath(
-                key.to_unquoted_string_str()?
-                    .split('.')
-                    .map(|s| s.to_owned())
-                    .collect(),
-            );
-            let item = SearchSpaceItem::new(value)?;
-            items.insert(path, item);
-        }
-        Ok(Self { items })
-    }
-
-    fn to_ask_search_space(&self) -> String {
-        nojson::json(|f| {
-            f.object(|f| {
-                for (path, item) in &self.items {
-                    f.member(
-                        path.0.join("."),
-                        nojson::json(|f| item.to_optuna_distribution(f)),
-                    )?;
-                }
-                Ok(())
-            })
-        })
-        .to_string()
-    }
-}
-
-#[derive(Debug)]
-enum SearchSpaceItem {
-    Number { min: JsonNumber, max: JsonNumber },
-    Categorical(Vec<JsonValue>),
-}
-
-impl SearchSpaceItem {
-    fn new(value: nojson::RawJsonValue<'_, '_>) -> Result<Self, nojson::JsonParseError> {
-        if value.kind().is_array() {
-            Ok(Self::Categorical(value.try_into()?))
-        } else if let Ok(object) = JsonObject::new(value) {
-            Ok(Self::Number {
-                min: object.get_required("min")?,
-                max: object.get_required("max")?,
-            })
-        } else {
-            Err(value.invalid("not JSON array or JSON object"))
-        }
-    }
-
-    fn to_optuna_distribution(&self, f: &mut nojson::JsonFormatter<'_, '_>) -> std::fmt::Result {
-        match self {
-            SearchSpaceItem::Number {
-                min: JsonNumber::Integer(min),
-                max: JsonNumber::Integer(max),
-            } => {
-                // 両方整数なら IntDistribution
-                f.object(|f| {
-                    f.member("name", "IntDistribution")?;
-                    f.member(
-                        "attributes",
-                        nojson::json(|f| {
-                            f.object(|f| {
-                                f.member("low", min)?;
-                                f.member("high", max)
-                            })
-                        }),
-                    )
-                })?;
-            }
-            SearchSpaceItem::Number { min, max } => {
-                // それ以外の数値は FloatDistribution
-                f.object(|f| {
-                    f.member("name", "FloatDistribution")?;
-                    f.member(
-                        "attributes",
-                        nojson::json(|f| {
-                            f.object(|f| {
-                                f.member("low", min)?;
-                                f.member("high", max)
-                            })
-                        }),
-                    )
-                })?;
-            }
-            SearchSpaceItem::Categorical(choices) => {
-                f.object(|f| {
-                    f.member("name", "CategoricalDistribution")?;
-                    f.member(
-                        "attributes",
-                        nojson::json(|f| f.object(|f| f.member("choices", choices))),
-                    )
-                })?;
-            }
-        }
-        Ok(())
-    }
-}
-
-fn check_optuna_availability() -> orfail::Result<()> {
-    let output = Command::new("optuna")
-        .arg("--version")
-        .stdout(Stdio::null())
-        .stderr(Stdio::null())
-        .output();
-
-    match output {
-        Ok(output) if output.status.success() => Ok(()),
-        Ok(_) => Err(orfail::Failure::new(
-            "optuna command failed to execute properly",
-        )),
-        Err(e) if e.kind() == std::io::ErrorKind::NotFound => Err(orfail::Failure::new(
-            "optuna command not found. Please install optuna and ensure it's in your PATH",
-        )),
-        Err(e) => Err(orfail::Failure::new(format!(
-            "failed to check optuna availability: {e}"
-        ))),
-    }
-}
-
-#[derive(Debug)]
-struct Optuna {
-    study_name: String,
-    storage_url: String,
-}
-
-impl Optuna {
-    fn new(study_name: String, storage_url: String) -> Self {
-        Self {
-            study_name,
-            storage_url,
-        }
-    }
-
-    fn create_study(&self) -> orfail::Result<()> {
-        let output = Command::new("optuna")
-            .arg("create-study")
-            .arg("--study-name")
-            .arg(&self.study_name)
-            .arg("--storage")
-            .arg(&self.storage_url)
-            .arg("--skip-if-exists")
-            // 「エンコード効率（何倍速変換か）の最小化」と「VMAF スコアの最大化」
-            .arg("--directions")
-            .arg("maximize")
-            .arg("maximize")
-            .stdout(Stdio::inherit())
-            .stderr(Stdio::inherit())
-            .output()
-            .or_fail_with(|e| format!("failed to execute optuna create-study command: {e}"))?;
-
-        output.status.success().or_fail()?;
-        Ok(())
-    }
-
-    fn ask(&self, search_space: &SearchSpace) -> orfail::Result<AskOutput> {
-        let search_space_json = search_space.to_ask_search_space();
-        log::debug!("ask search space: {search_space_json}");
-
-        // Execute optuna ask command
-        let output = Command::new("optuna")
-            .arg("ask")
-            .arg("--storage")
-            .arg(&self.storage_url)
-            .arg("--study-name")
-            .arg(&self.study_name)
-            .arg("--search-space")
-            .arg(&search_space_json)
-            .stdout(Stdio::piped())
-            .stderr(Stdio::piped())
-            .output()
-            .or_fail_with(|e| format!("failed to execute optuna ask command: {e}"))?;
-
-        output.status.success().or_fail_with(|()| {
-            let stderr = String::from_utf8_lossy(&output.stderr);
-            format!("optuna ask command failed: {stderr}")
-        })?;
-
-        // Parse the output to extract parameters
-        let stdout = String::from_utf8(output.stdout).or_fail()?;
-        let output = stdout.parse().map(|nojson::Json(v)| v).or_fail()?;
-        log::info!("ask result: {output:?}");
-
-        Ok(output)
-    }
-
-    fn tell(&self, trial_number: usize, metrics: &TrialMetrics) -> orfail::Result<()> {
-        let output = Command::new("optuna")
-            .arg("tell")
-            .arg("--storage")
-            .arg(&self.storage_url)
-            .arg("--study-name")
-            .arg(&self.study_name)
-            .arg("--trial-number")
-            .arg(trial_number.to_string())
-            .arg("--values")
-            .arg(&metrics.encoding_speed_ratio.to_string())
-            .arg(&metrics.vmaf_mean.to_string())
-            .arg("--state")
-            .arg("complete")
-            // optuna tell コマンドは「実験的機能です」という警告を出すけど、
-            // Hisui 側で対処できるものでもなく、ノイジーなだけなので抑制する
-            .env("PYTHONWARNINGS", "ignore")
-            .stdout(Stdio::inherit())
-            .stderr(Stdio::inherit())
-            .output()
-            .or_fail_with(|e| format!("failed to execute optuna tell command: {e}"))?;
-
-        output
-            .status
-            .success()
-            .or_fail_with(|()| "optuna tell command failed".to_string())?;
-
-        Ok(())
-    }
-
-    fn tell_fail(&self, trial_number: usize) -> orfail::Result<()> {
-        let output = Command::new("optuna")
-            .arg("tell")
-            .arg("--storage")
-            .arg(&self.storage_url)
-            .arg("--study-name")
-            .arg(&self.study_name)
-            .arg("--trial-number")
-            .arg(trial_number.to_string())
-            .arg("--state")
-            .arg("fail")
-            // optuna tell コマンドは「実験的機能です」という警告を出すけど、
-            // Hisui 側で対処できるものでもなく、ノイジーなだけなので抑制する
-            .env("PYTHONWARNINGS", "ignore")
-            .stdout(Stdio::inherit())
-            .stderr(Stdio::inherit())
-            .output()
-            .or_fail_with(|e| format!("failed to execute optuna tell command: {e}"))?;
-
-        output
-            .status
-            .success()
-            .or_fail_with(|()| "optuna tell command failed".to_string())?;
-
-        Ok(())
-    }
-}
-
-#[derive(Debug)]
-struct AskOutput {
-    trial_number: usize,
-    params: BTreeMap<JsonObjectMemberPath, JsonValue>,
-}
-
-impl AskOutput {
-    fn update_layout(&self, layout: &mut JsonValue) -> orfail::Result<()> {
-        for (path, new_value) in &self.params {
-            *path.get_mut(layout).or_fail()? = new_value.clone();
-        }
-        Ok(())
-    }
-}
-
-impl<'text, 'raw> TryFrom<nojson::RawJsonValue<'text, 'raw>> for AskOutput {
-    type Error = nojson::JsonParseError;
-
-    fn try_from(value: nojson::RawJsonValue<'text, 'raw>) -> Result<Self, Self::Error> {
-        Ok(Self {
-            trial_number: value.to_member("number")?.required()?.try_into()?,
-            params: value.to_member("params")?.required()?.try_into()?,
-        })
-    }
 }
 
 fn run_trial_evaluation(
@@ -607,10 +297,4 @@ fn run_trial_evaluation(
         encoding_speed_ratio,
         vmaf_mean,
     })
-}
-
-#[derive(Debug)]
-struct TrialMetrics {
-    encoding_speed_ratio: f64,
-    vmaf_mean: f64,
 }
